add translate to bib cli
