+++ pyproject.toml
@@ -29,4 +29,4 @@ requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
-evid = "evid.cli:main"
+evid = "evid.cli.main:main"
