+++ src/evid/cli/__init__.py
@@ -1,0 +1,1 @@
+# Empty __init__.py to make cli a module
